+++ config/symbols.us.txt
@@ -895,6 +895,12 @@ PLAYER_ext_player_anim = 0x80073484;
 g_Entities_1 = 0x80073494;
 g_Entities_19 = 0x800741CC;
 g_Entities_64 = 0x800762D8;
+DOPPLEGANGER_posX_i_hi = 0x800762DA;
+DOPPLEGANGER_posY_i_hi = 0x800762DE;
+DOPPLEGANGER_hitboxOffX = 0x800762E8;
+DOPPLEGANGER_hitboxOffY = 0x800762EA;
+DOPPLEGANGER_facingLeft = 0x800762EC;
+DOPPLEGANGER_animSet = 0x8007632C;
 g_Entities_160 = 0x8007A958;
 g_Entities_224 = 0x8007D858;
 g_EvHwCardEnd = 0x8007EFD8;
@@ -982,9 +988,3 @@ g_DebugPlayer = 0x80098850;
 _svm_brr_start_addr = 0x80098854;
 _svm_okof1 = 0x80098898;
 _svm_okof2 = 0x8009889C;
-DOPPLEGANGER_posX_i_hi = 0x800762DA;
-DOPPLEGANGER_posY_i_hi = 0x800762DE;
-DOPPLEGANGER_hitboxOffX = 0x800762E8;
-DOPPLEGANGER_hitboxOffY = 0x800762EA;
-DOPPLEGANGER_facingLeft = 0x800762EC;
-DOPPLEGANGER_animSet = 0x8007632C;
