+++ src/decelerate.h
@@ -1,0 +1,29 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+void DecelerateX(s32 amount) {
+    if (g_CurrentEntity->velocityX < 0) {
+        g_CurrentEntity->velocityX += amount;
+        if (g_CurrentEntity->velocityX > 0) {
+            g_CurrentEntity->velocityX = 0;
+        }
+    } else {
+        g_CurrentEntity->velocityX -= amount;
+        if (g_CurrentEntity->velocityX < 0) {
+            g_CurrentEntity->velocityX = 0;
+        }
+    }
+}
+
+void DecelerateY(s32 amount) {
+    if (g_CurrentEntity->velocityY < 0) {
+        g_CurrentEntity->velocityY += amount;
+        if (g_CurrentEntity->velocityY > 0) {
+            g_CurrentEntity->velocityY = 0;
+        }
+    } else {
+        g_CurrentEntity->velocityY -= amount;
+        if (g_CurrentEntity->velocityY < 0) {
+            g_CurrentEntity->velocityY = 0;
+        }
+    }
+}
