+++ src/dra/7879C.c
@@ -8,34 +8,7 @@
 #include "dra_bss.h"
 #include "servant.h"
 
-// Given a range of start and end values, finds an unused entity
-// slot in g_Entities to fill in. Starts at start, and goes to
-// end-1. If none in the range are available, returns NULL.
-Entity* GetFreeEntity(s16 start, s16 end) {
-    Entity* entity = &g_Entities[start];
-    s16 i;
-
-    for (i = start; i < end; i++, entity++) {
-        if (entity->entityId == E_NONE) {
-            return entity;
-        }
-    }
-    return NULL;
-}
-
-// Like GetFreeEntity, but searches for an open spot
-// starting from the end and working backward
-Entity* GetFreeEntityReverse(s16 start, s16 end) {
-    Entity* entity = &g_Entities[end - 1];
-    s16 i;
-
-    for (i = end - 1; i >= start; i--, entity--) {
-        if (entity->entityId == E_NONE) {
-            return entity;
-        }
-    }
-    return NULL;
-}
+#include "../get_free_entity.h"
 
 void func_80118894(Entity* self) {
     s32 i;
