+++ src/dra/84B88.c
@@ -1401,27 +1401,7 @@ void func_80127CC8(Entity* self) {
     }
 }
 
-// These are a pair of helper functions used for the rebound stone rebounding.
-// No clear reason why the first one puts the first line outside the
-// if statement.
-
-static void ReboundStoneBounce1(s16 bounceAngle) {
-    g_CurrentEntity->ext.reboundStone.stoneAngle =
-        (bounceAngle * 2) - g_CurrentEntity->ext.reboundStone.stoneAngle;
-    if (g_CurrentEntity->ext.reboundStone.unk82 == 0) {
-        g_CurrentEntity->ext.reboundStone.unk80++;
-        g_CurrentEntity->ext.reboundStone.unk82++;
-    }
-}
-
-static void ReboundStoneBounce2(s16 bounceAngle) {
-    if (g_CurrentEntity->ext.reboundStone.unk82 == 0) {
-        g_CurrentEntity->ext.reboundStone.stoneAngle =
-            (bounceAngle * 2) - g_CurrentEntity->ext.reboundStone.stoneAngle;
-        g_CurrentEntity->ext.reboundStone.unk80++;
-        g_CurrentEntity->ext.reboundStone.unk82++;
-    }
-}
+#include "../rebound_stone.h"
 
 // Entity ID 20. Created by blueprint 24. This comes from BlueprintNum for
 // the rebound stone SubweaponDef.
